// webapp/src/app/core/constants/query-param-operator.enum.ts
export enum QueryParamOperator {
	EQUALS = "eq",
	EQUALS_IGNORE_CASE = "eqIc",
	NOT_EQUAL = "neq",
	NOT_EQUAL_IGNORE_CASE = "neqIc",
	LIKE = "like",
	LIKE_IGNORE_CASE = "likeIc",
	NOT_LIKE = "notLike",
	NOT_LIKE_IGNORE_CASE = "notLikeIc",
	STARTS_WITH = "sw",
	STARTS_WITH_IGNORE_CASE = "swIc",
	ENDS_WITH = "ew",
	ENDS_WITH_IGNORE_CASE = "ewIc",
	IN = "in",
	IN_IGNORE_CASE = "inIc",
	NOT_IN = "ni",
	NOT_IN_IGNORE_CASE = "niIc",
	GREATER_THAN = "gt",
	LESS_THAN = "lt",
	GREATER_THAN_OR_EQUAL = "gte",
	LESS_THAN_OR_EQUAL = "lte",
	BETWEEN = "between",

	SPECIAL_FILTER_KEY = "specialFilter",
	IN_SEPARATOR = ",",
	KEY_SEPARATOR = ".",
	AND_SEPARATOR = "[and]",
	AND_SEPARATOR_GROUP = "[[and]]",
	OR_SEPARATOR = "[or]",
	OR_SEPARATOR_GROUP = "[[or]]",
}